Update log_rotation_delay to be a DURATION
Compatible with old values since default unit is seconds,
and DURATION takes Long values whereas the setting
previously was INTEGERS.

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/factory/PlatformModule.java
@@ -216,7 +216,7 @@ public class PlatformModule
     protected LogService createLogService( LogProvider userLogProvider )
     {
         long internalLogRotationThreshold = config.get( GraphDatabaseSettings.store_internal_log_rotation_threshold );
-        int internalLogRotationDelay = config.get( GraphDatabaseSettings.store_internal_log_rotation_delay );
+        long internalLogRotationDelay = config.get( GraphDatabaseSettings.store_internal_log_rotation_delay );
         int internalLogMaxArchives = config.get( GraphDatabaseSettings.store_internal_log_max_archives );
 
         final StoreLogService.Builder builder =
